CyanManager/requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
comtypes==1.4.15
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
keyboard==0.13.5
keyring==25.7.0
more-itertools==10.8.0
numpy==2.4.2
psutil==7.2.2
pycaw==20251023
pycparser==3.0
pygame==2.6.1
PyMonCtl==0.92
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.18.0
pyserial==3.5
python-dotenv==1.2.1
pywin32==311
pywin32-ctypes==0.2.3
PyWinBox==0.7
PyWinCtl==0.4.1
requests==2.32.5
screeninfo==0.8.1
sounddevice==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
WMI==1.5.1
openrgb-python
